refactor(web): the locale-independent formatters live once in web-shared/
formatCompact, formatTokens, formatDollars and formatKwh had identical bodies
in both frontends. They are the four that pin themselves to en-US rather than
the active language, which is exactly why they can be shared: everything else
in either format.ts phrases itself through i18next and stays where it is.

Each app re-exports them from its own utils/format, so no call site changes and
the modules keep one import path.

The tests follow the code. web/ owns web-shared/ for coverage, so it gains the
dollars, kWh and billions cases Front Desk had; Front Desk's suite narrows to
formatCount, the one formatter it still owns, which had no test before.
